+++ src/gameOverStage.h
@@ -24,7 +24,7 @@ public:
 
 	void update(float dt);
 
-	void resizeOptions(int width, int height);
+	void resizeOptions(float width, float height);
 	void changeOption(int to_add);
 	bool selectOption();
