+++ src/stage.h
@@ -19,7 +19,7 @@ public:
 	virtual void onExit() {};
 	virtual void render() {};
 	virtual void update(float dt) {};
-	virtual void resizeOptions(int width, int height) {};
+	virtual void resizeOptions(float width, float height) {};
 	void renderHealthBar(Vector3 position, float hp_ratio, Shader* shader, float width, float height);
 	void renderHungerBar(Vector3 position, float hunger_ratio, Shader* shader, float width, float height);
